requirements.txt
selenium
Appium-Python-Client
PyYAML>=5.1.1
PyMySQL
pymongo
lettuce